+++ src/extension.ts
@@ -11,6 +11,7 @@ import { ALFiles } from './al/alFiles';
 import { ReportCreator } from './al/report/reportCreator';
 import { variableRemover as VariableRemover } from './al/report/variableRemover';
 import * as LaunchJsonUpdater from './json/launchjson_updater';
+import * as XlfUpdater from './xlf/xlf_updater';
 import { ALCodeActionsProvider } from './al/codeActions/alCodeActionsProvider';
 const fieldHover = require('./additional/fieldHover');
 
@@ -67,6 +68,10 @@ export function activate(context: vscode.ExtensionContext) {
 		LaunchJsonUpdater.selectStartupObjectId();
 	});
 
+	let insertTranslationFromCommentCmd = commands.registerCommand("extension.insertTranslationFromComment", () => {
+		XlfUpdater.insertTranslationFromComment();
+	});
+
 	let translateAndCopyToClipboardCmd = commands.registerCommand("extension.TranslateAndCopyToClipboard", () => {
 		Translator.translateAndCopyToClipboard(false);
 	});
@@ -94,6 +99,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(translateAndCopyToClipboardCmd);
 	context.subscriptions.push(removeUnusedVarsInReportCmd);
 	context.subscriptions.push(selectStartupObjectIdCmd);
+	context.subscriptions.push(insertTranslationFromCommentCmd);
 }
 
 // this method is called when your extension is deactivated